fix: 認証回り全体の修正

+++ internal/auth/core/dto/dog_owner_dto.go
@@ -1,5 +1,5 @@
 package dto
 
 type ResDogOwnerDto struct {
-	DogOwnerID uint `json:"dogOwnerId"`
+	DogOwnerID uint64 `json:"dogOwnerId"`
 }
